requirements.txt
# 核心依赖
pyyaml>=6.0
pydantic>=2.0
jinja2>=3.1.0
requests>=2.28.0  # For Ollama client HTTP requests

# 命令行界面
click>=8.0

# 开发和测试
pytest>=7.0
pytest-cov>=4.0
black>=23.0
ruff>=0.1.0
mypy>=1.0

# 可选：分子结构处理
# numpy>=1.24.0  # 如果需要坐标处理

# AI 功能依赖（可选）
openai>=1.0.0          # OpenAI API 支持
anthropic>=0.3.0       # Anthropic Claude API 支持
ollama>=0.1.0          # Ollama 本地模型支持（可选）
# langchain>=0.1.0     # 可选：LangChain 支持（用于复杂 AI 工作流）

# 数据标准化依赖
jsonschema>=4.0.0      # JSON Schema 验证
jsonlines>=3.0.0       # JSONL 文件处理
pandas>=1.5.0          # 可选：用于 Parquet 格式导出

